feat: simplify auth check - only on API write actions
